refactor classes a bit
