+++ dapps/marketplace/src/pages/listings/Listings.js
@@ -224,7 +224,7 @@ class Listings extends Component {
 export default withGrowthCampaign(
   withWallet(withTokenBalance(withCreatorConfig(Listings))),
   {
-    useCache: true,
+    fetchPolicy: 'cache-first',
     queryEvenIfNotEnrolled: true,
     suppressErrors: true // still show listings in case growth can not be reached
   }
