+++ dapps/marketplace/src/utils/growthTools.js
@@ -31,33 +31,3 @@ export function getAttestationReward({
     return ''
   }
 }
-
-export function getGrowthListingsRewards({ growthCampaigns, tokenDecimals }) {
-  const ognListingRewards = {}
-
-  if (growthCampaigns && tokenDecimals) {
-    const activeCampaign = growthCampaigns.find(
-      campaign => campaign.status === 'Active'
-    )
-
-    if (!activeCampaign) return ognListingRewards
-
-    const decimalDivision = web3.utils
-      .toBN(10)
-      .pow(web3.utils.toBN(tokenDecimals))
-
-    activeCampaign.actions.forEach(action => {
-      if (action.type === 'ListingIdPurchased') {
-        const normalisedReward = parseInt(
-          web3.utils
-            .toBN(action.reward ? action.reward.amount : 0)
-            .div(decimalDivision)
-            .toString()
-        )
-        ognListingRewards[action.listingId] = normalisedReward
-      }
-    })
-  }
-
-  return ognListingRewards
-}
